Fix windows test multiple definitions
